[SLS-12] Updated spring interceptor
