+++ assertion-rest-client-native/build.gradle
@@ -10,6 +10,7 @@
 plugins {
 	// Apply the java-library plugin for API and implementation separation.
 	id 'java-library'
+	id "java-test-fixtures"
 	id("io.freefair.lombok") version "8.0.0"
 /*id("org.openapi.generator") version "6.5.0"*/
 }
