+++ spring-impl/build.gradle
@@ -22,8 +22,10 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-web:2.7.10'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test:2.7.10'
 	testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.2'
+	//Mockserver for testing api
+	testImplementation 'org.mock-server:mockserver-netty:5.15.0'
 	testImplementation project(':http-verifier')
-	testImplementation project(':assertion-rest-client-native')
+	testImplementation testFixtures(project(':assertion-rest-client-native'))
 	testImplementation project(':identity-service-rest-client-native')
 }
