+++ spring-impl/src/test/java/it/pagopa/tech/lollipop/consumer/spring/DemoApplication.java
@@ -1,0 +1,13 @@
+/* (C)2023 */
+package it.pagopa.tech.lollipop.consumer.spring;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = "it.pagopa.tech")
+public class DemoApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DemoApplication.class, args);
+    }
+}
