+++ spring-impl/src/test/java/it/pagopa/tech/lollipop/consumer/spring/HttpVerifierHandlerInterceptorIntegrationTest.java
@@ -1,114 +1,96 @@
 /* (C)2023 */
 package it.pagopa.tech.lollipop.consumer.spring;
 
-import it.pagopa.tech.lollipop.consumer.assertion.AssertionServiceFactory;
-import it.pagopa.tech.lollipop.consumer.assertion.client.simple.AssertionSimpleClientConfig;
-import it.pagopa.tech.lollipop.consumer.assertion.client.simple.AssertionSimpleClientProvider;
-import it.pagopa.tech.lollipop.consumer.assertion.impl.AssertionServiceFactoryImpl;
-import it.pagopa.tech.lollipop.consumer.assertion.storage.SimpleAssertionStorageProvider;
-import it.pagopa.tech.lollipop.consumer.assertion.storage.StorageConfig;
-import it.pagopa.tech.lollipop.consumer.http_verifier.HttpMessageVerifierFactory;
-import it.pagopa.tech.lollipop.consumer.http_verifier.visma.VismaHttpMessageVerifierFactory;
-import it.pagopa.tech.lollipop.consumer.idp.IdpCertProviderFactory;
-import it.pagopa.tech.lollipop.consumer.idp.impl.IdpCertProviderFactoryImplStub;
+import static org.mockserver.integration.ClientAndServer.startClientAndServer;
+
 import it.pagopa.tech.lollipop.consumer.spring.config.HttpVerifierConfiguration;
 import it.pagopa.tech.lollipop.consumer.spring.config.SpringLollipopConsumerRequestConfig;
-import java.security.*;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockserver.integration.ClientAndServer;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.stereotype.Component;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.web.filter.CommonsRequestLoggingFilter;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.client.RestTemplate;
 
 @ExtendWith(SpringExtension.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ContextConfiguration(
         classes = {
-            HttpVerifierHandlerInterceptor.class,
+            DemoApplication.class,
             SpringLollipopConsumerRequestConfig.class,
             HttpVerifierConfiguration.class,
-            HttpVerifierHandlerInterceptorIntegrationTest.DemoServicesConfig.class
         })
 public class HttpVerifierHandlerInterceptorIntegrationTest {
 
-    //    @Autowired
-    //    private DemoController controller;
-    //
-    //    @LocalServerPort
-    //    private int port;
-    //
-    //    @Controller
-    //    public static class DemoController {
-    //
-    //        @PostMapping("/")
-    //        public @ResponseBody
-    //        String testMethod() {
-    //            return "Test!";
-    //        }
-    //
-    //    }
+    @LocalServerPort private int port;
+    @Autowired private TestRestTemplate restTemplate;
+
+    @BeforeAll
+    public static void startServer() {
+        mockServer = startClientAndServer(3000);
+    }
+
+    private static ClientAndServer mockServer;
 
     @Autowired private HttpVerifierHandlerInterceptor interceptor;
 
-    @Configuration
-    public static class DemoServicesConfig {
-
-        @Bean
-        public SpringLollipopConsumerRequestConfig verifierConfiguration() {
-            return new SpringLollipopConsumerRequestConfig();
-        }
-
-        @Bean
-        public HttpMessageVerifierFactory httpMessageVerifierFactory() throws Exception {
-            return new VismaHttpMessageVerifierFactory("UTF-8", verifierConfiguration());
-        }
-
-        @Bean
-        public IdpCertProviderFactory idpCertProviderFactory() {
-            return new IdpCertProviderFactoryImplStub();
-        }
-
-        @Bean
-        public AssertionServiceFactory assertionServiceFactory() {
-            return new AssertionServiceFactoryImpl(
-                    new SimpleAssertionStorageProvider(),
-                    new AssertionSimpleClientProvider(
-                            AssertionSimpleClientConfig.builder().build()),
-                    storageConfig());
-        }
-
-        @Bean
-        public StorageConfig storageConfig() {
-            return new StorageConfig();
-        }
-    }
-
-    @Component
-    public static class DemoWebConfigurer implements WebMvcConfigurer {
-
-        @Autowired private HttpVerifierHandlerInterceptor interceptor;
-
-        @Bean
-        public CommonsRequestLoggingFilter loggingFilter() {
-            CommonsRequestLoggingFilter filter = new CommonsRequestLoggingFilter();
-            filter.setIncludeQueryString(true);
-            filter.setIncludeClientInfo(true);
-            filter.setIncludeHeaders(true);
-            filter.setIncludePayload(true);
-            return filter;
-        }
-
-        @Override
-        public void addInterceptors(InterceptorRegistry registry) {
-            registry.addInterceptor(interceptor);
-        }
-    }
-
     @Test
-    public void something() {}
+    public void test() {
+        AssertionSimpleClientTestUtils.createExpectationAssertionFound();
+
+        String signatureInput =
+                "sig123=(\"content-digest\" \"x-pagopa-lollipop-original-method\""
+                    + " \"x-pagopa-lollipop-original-url\");created=1678293988;nonce=\"aNonce\";alg=\"ecdsa-p256-sha256\";keyid=\"sha256-a7qE0Y0DyqeOFFREIQSLKfu5WlbckdxVXKFasfcI-Dg\"";
+        var signature =
+                "sig123=:lTuoRytp53GuUMOB4Rz1z97Y96gfSeEOm/xVpO39d3HR6lLAy4KYiGq+1hZ7nmRFBt2bASWEpen7ov5O4wU3kQ==:";
+
+        RestTemplate exec = restTemplate.getRestTemplate();
+        exec.getClientHttpRequestInitializers()
+                .add(
+                        request -> {
+                            request.getHeaders()
+                                    .add(
+                                            "Content-Digest",
+                                            "sha-256=:cpyRqJ1VhoVC+MSs9fq4/4wXs4c46EyEFriskys43Zw=:");
+                            request.getHeaders()
+                                    .add(
+                                            "x-pagopa-lollipop-original-url",
+                                            "https://api-app.io.pagopa.it/first-lollipop/sign");
+                            request.getHeaders().add("x-pagopa-lollipop-original-method", "POST");
+                            request.getHeaders()
+                                    .add(
+                                            "x-pagopa-lollipop-public-key",
+                                            "eyJrdHkiOiJFQyIsIngiOiJGcUZEd"
+                                                + "XdFZ3U0TVVYRVJQTVZMLTg1cEd2MkQzWW1MNEoxZ2ZNa2RiYzI0IiwieSI6Im"
+                                                + "hkVjBveG1XRlN4TW9KVURwZGlocjc2clM4VlJCRXFNRmViWXlBZks5LWsiLCJjcnYiO"
+                                                + "iJQLTI1NiJ9");
+                            request.getHeaders()
+                                    .add(
+                                            "x-pagopa-lollipop-assertion-ref",
+                                            "sha256-_ZzL8qeuAM5kQ9pbMB4tn7IDSQZCVXAkW9fm4P7ULPI");
+                            request.getHeaders().add("x-pagopa-lollipop-assertion-type", "SAML");
+                            request.getHeaders().add("x-pagopa-lollipop-auth-jwt", "aValidJWT");
+                            request.getHeaders()
+                                    .add("x-pagopa-lollipop-user-id", "AAAAAA89S20I111X");
+                            request.getHeaders().add("Signature-Input", signatureInput);
+                            request.getHeaders().add("Signature", signature);
+                        });
+
+        String response =
+                exec.postForObject(
+                        "http://localhost:" + port,
+                        "{\"message\":\"a valid message payload\"}",
+                        String.class);
+    }
+
+    @AfterAll
+    public static void stopServer() {
+        mockServer.stop();
+    }
 }
