+++ spring-impl/src/test/java/it/pagopa/tech/lollipop/consumer/spring/config/DemoWebConfigurer.java
@@ -1,0 +1,31 @@
+/* (C)2023 */
+package it.pagopa.tech.lollipop.consumer.spring.config;
+
+import it.pagopa.tech.lollipop.consumer.spring.HttpVerifierHandlerInterceptor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.CommonsRequestLoggingFilter;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Component
+public class DemoWebConfigurer implements WebMvcConfigurer {
+
+    @Autowired private HttpVerifierHandlerInterceptor interceptor;
+
+    @Bean
+    public CommonsRequestLoggingFilter loggingFilter() {
+        CommonsRequestLoggingFilter filter = new CommonsRequestLoggingFilter();
+        filter.setIncludeQueryString(true);
+        filter.setIncludeClientInfo(true);
+        filter.setIncludeHeaders(true);
+        filter.setIncludePayload(true);
+        return filter;
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(interceptor);
+    }
+}
